Change product price type from double to int
Updated the Product model and related logic to use int for price instead of double, ensuring consistency in data handling. Adjusted parsing and currency formatting to accommodate the new type.

+++ lib/models/product.dart
@@ -17,7 +17,7 @@ class Product {
   String restaurantId;
   String id;
   String name;
-  double price;
+  int price;
   String categoryId;
   bool delete;
   String imageUrl;
@@ -41,7 +41,7 @@ class Product {
       restaurantId: data['restaurantId'] ?? '',
       id: data['id'] ?? '',
       name: data['name'] ?? '',
-      price: (data['price'] ?? 0) as double,
+      price: (data['price'] ?? 0) as int,
       categoryId: data['category_id'] ?? '',
       delete: data['delete'] ?? false,
       imageUrl: data['image_url'] ?? '',
